Refactored to move DLZ config into CommonInclude.
The snippet is also in Week2, but probably makes sense to keep it there since its after the initial setup. The snippet in Week5 needs to stay as is I believe because it refers to a different container. I also found a slight bug with a missing import for graph_objects and fixed that one. I ran through all the affected notebooks and confirmed they're working.